pyproject.toml
[project]
name = "syntropic137"
version = "0.3.0"
description = "Event-sourced system for tracking AI agent work across workflows"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "syntropic137" }]
keywords = ["ai", "agents", "event-sourcing", "workflows", "metrics"]

# Workspace member dependencies
dependencies = [
    "syn-api",
    "syn-cli",
    "syn-collector",
    "syn-dashboard",
    "syn-domain",
    "syn-adapters",
    "syn-shared",
    "event-sourcing-python",
    "grpcio>=1.60", # Required by event-sourcing SDK
    "minio>=7.2.20",
    "sqlmodel>=0.0.27",
]

# Optional dependencies
[project.optional-dependencies]
all = []

[dependency-groups]
dev = [
    "ruff>=0.14.7",
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "types-pyyaml>=6.0.12",
    "psycopg[binary]>=3.3.0",
    "asyncpg>=0.31.0",
    "testcontainers>=4.13.3",
    "pyright>=1.1.408",
]

[build-system]
requires = ["uv_build>=0.9.13"]
build-backend = "uv_build"

[tool.uv.workspace]
members = ["apps/*", "packages/*"]
exclude = ["apps/syn-dashboard-ui", "apps/syn-docs"]

[tool.uv.sources]
syn-api = { workspace = true }
syn-cli = { workspace = true }
syn-collector = { workspace = true }
syn-dashboard = { workspace = true }
syn-domain = { workspace = true }
syn-adapters = { workspace = true }
syn-shared = { workspace = true }
syn-perf = { workspace = true }
syn-tokens = { workspace = true }
# Submodule dependencies
event-sourcing-python = { path = "lib/event-sourcing-platform/event-sourcing/python" }
agentic-logging = { path = "lib/agentic-primitives/lib/python/agentic_logging" }
agentic-isolation = { path = "lib/agentic-primitives/lib/python/agentic_isolation" }
agentic-events = { path = "lib/agentic-primitives/lib/python/agentic_events" }

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["apps/*/src", "packages/*/src"]
exclude = [
    "lib/",
    ".venv/",
    "__pycache__/",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
# Test files often have unused args in mocks implementing interfaces
"**/test_*.py" = ["ARG001", "ARG002"]
"**/tests/**" = ["ARG001", "ARG002"]
# Scripts may have planned-but-not-implemented params
"scripts/*.py" = ["ARG001"]
# API stub functions have complete signatures but no implementation
"**/syn_api/v1/artifacts.py" = ["ARG001"]
"**/syn_api/v1/github.py" = ["ARG001"]
"**/syn_api/v1/observability.py" = ["ARG001"]
# gRPC handlers require context param even if unused
"**/token_injector.py" = ["ARG002"]

[tool.ruff.lint.isort]
known-first-party = ["syn_api", "syn_cli", "syn_collector", "syn_dashboard", "syn_domain", "syn_adapters", "syn_shared", "syn_perf"]

[tool.mypy]
python_version = "3.12"
# Strict mode with pragmatic exceptions for existing code
strict = false  # Disabled until existing type debt is resolved
warn_return_any = false  # Too many violations in projections
warn_unused_ignores = false  # Type ignores may vary between envs (local vs CI)
disallow_untyped_defs = false  # TODO: Enable after fixing 26+ violations
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false  # Decorators often lack types
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true
pretty = true
# Exclude test files from type checking (both /tests/ dirs and co-located test_*.py)
exclude = [
    "^.*/tests/.*$",
    "^.*test_.*\\.py$",
]

# IMPORTANT: NEVER use ignore_errors = true in mypy overrides!
# Always fix type errors properly with:
#   - Proper type annotations
#   - cast() for legitimate type narrowing
#   - Guards (if x is not None, isinstance, etc.)
#
# The # type: ignore comment is also discouraged - use only as last resort
# with a specific error code and explanation.

# Per-module ignores for optional dependencies and internal modules
[[tool.mypy.overrides]]
module = [
    "redis.*",
    "aiohttp.*",
    "fakeredis.*",
    "syn_api.*",
    "syn_adapters.*",  # Internal module cross-references
    "syn_tokens.*",
    "syn_shared.*",
    "syn_domain.*",
]
ignore_missing_imports = true

# Ignore missing imports for optional dependencies
[[tool.mypy.overrides]]
module = [
    "anthropic",
    "anthropic.*",
    "openai",
    "openai.*",
    "asyncpg",
    "asyncpg.*",
    "agentic_logging",
    "agentic_logging.*",
    "agentic_isolation",
    "agentic_isolation.*",
    "agentic_events",
    "agentic_events.*",
    "sse_starlette",
    "sse_starlette.*",
    "uvicorn",
    "uvicorn.*",
    "grpc",
    "grpc.*",
    "supabase",
    "supabase.*",
    "minio",
    "minio.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "apps/syn-cli/tests",
    "apps/syn-dashboard/tests",
    "packages/syn-api/tests",
    "packages/syn-collector/tests",
    "packages/syn-domain/tests",
    "packages/syn-domain/src",  # VSA co-located tests
    "packages/syn-adapters/tests",
    "packages/syn-shared/tests",
]
pythonpath = [
    "apps/syn-cli/src",
    "apps/syn-dashboard/src",
    "packages/syn-api/src",
    "packages/syn-collector/src",
    "packages/syn-domain/src",
    "packages/syn-adapters/src",
    "packages/syn-shared/src",
]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = ["-ra", "-q", "--strict-markers", "-m", "not integration"]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may use Docker services)",
    "e2e: End-to-end tests (full system)",
    "regression: Regression tests for known bugs",
]

[tool.coverage.run]
branch = true
source = ["apps", "packages"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "raise typer.Exit",  # CLI exits are hard to test
]
# Exclude files that require external services or are deprecated
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/agents/claude.py",  # Requires Anthropic API
    "*/agents/openai.py",  # Requires OpenAI API
    "*/storage/postgres.py",  # Requires running PostgreSQL
    "*/workspace_backends/*",  # Requires Docker runtime
    "*/stale_execution_cleaner.py",  # Background cleanup - integration tested
]
